Prevent quick U turn
Add a new_dir field to snake struct. This will be assigned on
update to the current dir.

It was not enough to store a single direction, because users could
change the direction twice in 1 tick and therefore force a U turn,
since the validation code for the new direction would use the
currently stored direction (which was already updated that tick).

example. dir = right. update to up, update to left. This was
possible in 1 tick. The new version will allow update to up, but
not the second update to left.